pyproject.toml
[tool.poetry]
name = "claude-md-skills"
version = "0.1.0"
description = "Skills and utilities for Claude Code"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
# gemini-image-generator dependencies
google-genai = "^1.0.0"
python-dotenv = "^1.0.0"
# frontend-tester dependencies
playwright = "^1.57.0"
# webpage-to-markdown dependencies
requests = "^2.31.0"
pyyaml = "^6.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
